backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# RAG / Vector (MVP)
chromadb==0.5.23
sentence-transformers==3.3.1

# OCR (MVP)
pytesseract==0.3.13
pdf2image==1.17.0
PyPDF2==3.0.1

# DB (선택)
# asyncpg==0.30.0
# sqlalchemy[asyncio]==2.0.36

# WebSocket
websockets==14.1
